web_dashboard/requirements.txt
Flask>=2.3.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
scikit-image>=0.19.0
scipy>=1.8.0
requests>=2.28.0